🔧 reorganize plugin structure to use service provider
- delete mautic.php file as part of plugin refactor
- update mautic.xml to reflect changes in file structure and bump version to 3.0.0
- create new service provider for plugin registration
- adjust TokenField to use updated namespace for MauticApiHelper
