+++ plg_system_mautic/services/provider.php
@@ -1,0 +1,49 @@
+<?php
+
+/**
+ * @package     Mautic-Joomla.Plugin
+ * @subpackage  System.Mautic
+ * @author      Mautic
+ * @copyright   Copyright (C) 2014 - 2025 Mautic All Rights Reserved.
+ * @license     http://www.gnu.org/licenses/gpl-3.0.html GNU/GPL
+ * @link        http://www.mautic.org
+ */
+
+ \defined('_JEXEC') or die;
+
+ use Joomla\CMS\Extension\PluginInterface;
+ use Joomla\CMS\Factory;
+ use Joomla\CMS\Plugin\PluginHelper;
+ use Joomla\Database\DatabaseInterface;
+ use Joomla\DI\Container;
+ use Joomla\DI\ServiceProviderInterface;
+ use Joomla\Event\DispatcherInterface;
+ use Mautic\Plugin\System\Mautic\Extension\Mautic;
+
+ return new class () implements ServiceProviderInterface {
+    /**
+     * Registers the service provider with a DI container.
+     *
+     * @param   Container  $container  The DI container.
+     *
+     * @return  void
+     *
+     * @since   3.0.0
+     */
+    public function register(Container $container)
+    {
+        $container->set(
+            PluginInterface::class,
+            function (Container $container) {
+                $plugin     = new Mautic(
+                    $container->get(DispatcherInterface::class),
+                    (array) PluginHelper::getPlugin('system', 'mautic')
+                );
+                $plugin->setApplication(Factory::getApplication());
+                // $plugin->setDatabase($container->get(DatabaseInterface::class));
+
+                return $plugin;
+            }
+        );
+    }
+};
